+++ packages/backend/src/controllers/savedChartController.ts
@@ -1,4 +1,4 @@
-import { ApiErrorPayload } from '@lightdash/common';
+import { ApiErrorPayload, FiltersResponse } from '@lightdash/common';
 import { Body, Get, Post } from '@tsoa/runtime';
 import express from 'express';
 import {
@@ -14,7 +14,7 @@ import {
 } from 'tsoa';
 import { projectService } from '../services/services';
 import { allowApiKeyAuthentication, isAuthenticated } from './authentication';
-import { ApiRunQueryResponse, Filters } from './runQueryController';
+import { ApiRunQueryResponse } from './runQueryController';
 
 @Route('/api/v1/saved/{chartUuid}')
 @Response<ApiErrorPayload>('default', 'Error')
@@ -31,7 +31,7 @@ export class SavedChartController extends Controller {
     @Post('/results')
     @OperationId('postChartResults')
     async postDashboardTile(
-        @Body() body: { filters?: Filters },
+        @Body() body: { filters?: FiltersResponse },
         @Path() chartUuid: string,
         @Request() req: express.Request,
     ): Promise<ApiRunQueryResponse> {