+++ packages/backend/src/services/ProjectService/ProjectService.ts
@@ -490,7 +490,7 @@ export class ProjectService {
         await this.projectModel.update(projectUuid, updatedProject);
         await this.jobModel.create(job);
 
-        if (savedProject.dbtConnection.type !== DbtProjectType.NONE) {
+        if (updatedProject.dbtConnection.type !== DbtProjectType.NONE) {
             await schedulerClient.testAndCompileProject({
                 organizationUuid: user.organizationUuid,
                 createdByUserUuid: user.userUuid,