+++ packages/frontend/src/hooks/gdrive/useGdrive.ts
@@ -1,6 +1,12 @@
-import { ApiError, ApiGdriveAccessTokenResponse } from '@lightdash/common';
+import {
+    ApiError,
+    ApiGdriveAccessTokenResponse,
+    ApiScheduledDownloadCsv,
+    UploadMetricGsheet,
+} from '@lightdash/common';
 import { useQuery } from 'react-query';
 import { lightdashApi } from '../../api';
+import { convertDateFilters } from '../../utils/dateFilter';
 
 const getGdriveAccessToken = async () =>
     lightdashApi<ApiGdriveAccessTokenResponse['results']>({
@@ -15,3 +21,18 @@ export const useGdriveAccessToken = () =>
         queryFn: getGdriveAccessToken,
         retry: false,
     });
+
+export const uploadGsheet = async (gsheetMetric: UploadMetricGsheet) => {
+    const timezoneFixQuery = {
+        ...gsheetMetric.metricQuery,
+        filters: convertDateFilters(gsheetMetric.metricQuery.filters),
+    };
+    return lightdashApi<ApiScheduledDownloadCsv>({
+        url: `/gdrive/upload-gsheet`,
+        method: 'POST',
+        body: JSON.stringify({
+            ...gsheetMetric,
+            metricQuery: timezoneFixQuery,
+        }),
+    });
+};